streamlit_chatbot/requirements.txt
streamlit>=1.24.0
boto3>=1.28.0
Pillow>=9.0.0
python-dotenv>=1.0.0